+++ test/dime/core_test.clj
@@ -80,14 +80,15 @@
 
 (deftest test-dependency-graph
   (let [f (fn [id ds] (reify t/Injectable
-                        (valid?   [_] true)
-                        (id-key   [_] id)
-                        (dep-keys [_] ds)
-                        (inject   [_ deps pre] :injected)
-                        (pre-inject-fn  [_] nil)
-                        (post-inject-fn [_] nil)))
+                        (valid? [_] true)
+                        (iattrs [_] (t/map->InjectableAttributes {:inj-id  id
+                                                                  :impl-id  (gensym)
+                                                                  :dep-ids  ds
+                                                                  :pre-inj  nil
+                                                                  :post-inj nil}))
+                        (inject   [_ deps pre] :injected)))
         g (reduce (fn [m x]
-                    (assoc m (t/id-key x) x))
+                    (assoc m (.-inj-id (t/iattrs x)) x))
             {} [(f :foo [:bar :baz])
                 (f :bar [:x :y])
                 (f :baz [:p :q])